Extract to external class
